// htdocs/rescueman/src/app/shared/models/veterinary.ts
export class Veterinary {
    id: number;
    name: string;
    email: string;
    phone: string;
    veterinary_id: number;

    street_line_1: string;
    street_line_2: string;
    city: string;
    state_prov: string;
    postal_code: string;

    created_by: number;
    updated_by: number;
    statecode: string;
    created_at: Date;
    updated_at: Date;
}